refact(auth-design-integration) : Remove barrel pattern for component imports
