+++ src/components/common/Button/index.tsx
@@ -1,7 +1,6 @@
 'use client';
 import { cn } from '@/lib/utils';
 import { cva, type VariantProps } from 'class-variance-authority';
-import { MouseEventHandler } from 'react';
 
 export const buttonVariants = cva(
   "inline-flex items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-all disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4 shrink-0 [&_svg]:shrink-0 outline-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive",
@@ -39,10 +38,6 @@ export const Button = ({
 }: React.ComponentProps<'button'> &
   VariantProps<typeof buttonVariants> & {
     children: React.ReactNode;
-    type?: 'button' | 'submit' | 'reset';
-    variant?: VariantProps<typeof buttonVariants>;
-    size?: VariantProps<typeof buttonVariants>;
-    onClick?: MouseEventHandler<HTMLButtonElement>;
   }) => {
   return (
     <button
