--- src/components/common/ImageSlider/index.ts
@@ -1,1 +1,0 @@
-export * from './ImageSlider';
